+++ src/main/java/com/itemmania/service/boardService/Search/BoardSearchService.java
@@ -4,6 +4,8 @@ import com.itemmania.entity.BoardEntity;
 import com.itemmania.repository.BoardRepository;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,22 +18,22 @@ public class BoardSearchService {
     private BoardRepository boardRepository;
 
     //삽니다, 팝니다 구분하여 사용자가 입력한 게임or게임서버를 검색후 프리미엄 결과값만 출력 
-    public List<BoardEntity> getP_SearchDataCheck(String dealCheck, String gameName, String gameServerName) {
+    public Page<BoardEntity> getP_SearchDataCheck(Pageable pageable, String dealCheck, String gameName, String gameServerName) {
 
         log.info(dealCheck);
         log.info(gameName + gameServerName);
 
-        List<BoardEntity> P_list = boardRepository.boardP_SearchList(dealCheck, gameName, gameServerName);
+        Page<BoardEntity> P_list = boardRepository.boardP_SearchList(pageable, dealCheck, gameName, gameServerName);
         log.info("검색서비스_프리미엄" + P_list);
         return P_list;
     }
 
     //삽니다, 팝니다 구분하여 사용자가 입력한 게임or게임서버를 검색후 일반 결과값만 출력
-    public List<BoardEntity> getSearchDataCheck(String dealCheck, String gameName, String gameServerName) {
+    public List<BoardEntity> getSearchDataCheck(Pageable pageable, String dealCheck, String gameName, String gameServerName) {
         log.info(dealCheck);
         log.info(gameName + gameServerName);
 
-        List<BoardEntity> list = boardRepository.boardSearchList(dealCheck, gameName, gameServerName);
+        List<BoardEntity> list = boardRepository.boardSearchList(pageable, dealCheck, gameName, gameServerName);
         log.info("검색서비스_일반" + list);
         return list;
     }
