+++ src/main/resources/templates/fragment/nav.html
@@ -10,8 +10,8 @@
                 <li class="highlight"><a href="#">구매등록</a></li>
             </ul>
             <ul class="nav nav_ct">
-                <li><a href="/board/boardList?deal=sale">삽니다<i class="icon_new"></i></a></li>
-                <li><a href="/board/boardList?deal=buy">팝니다<i class="icon_new"></i></a></li>
+                <li><a href="/board/boardList?deal=buy">삽니다<i class="icon_new"></i></a></li>
+                <li><a href="/board/boardList?deal=sale">팝니다<i class="icon_new"></i></a></li>
                 <li><a href="/myroom">마이룸</a></li>
                 <li><a href="#">포인트 충전</a></li>
                 <li><a href="#"></a></li>
