+++ src/main/resources/templates/fragment/search.html
@@ -10,6 +10,9 @@
 
         console.log(st + "-" + sGameServer);
 
+
+        //검색시 링크에 deal=sale   deal=buy 값 파라미터로 넣어서 보내줘야함
+
     }
 </script>
 <div class="g_logo_area" xmlns:th="http://www.w3.org/1999/html">
@@ -21,7 +24,7 @@
     </div>
     </style>
     <div class="g_search_top">
-        <form id="g_searchbar_form" method="post" action="/boardSearch" onsubmit="return searchbarSubmit();">
+        <form id="g_searchbar_form" method="get" action="/board/boardList" onsubmit="return searchbarSubmit();">
             <div class="g_search_wrapper">
                 <div class="search_area">
                     <div class="g_trade_type">
